# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Aapi",
    "version": "1.0"
  },
  "paths": {
    "/api/Anime": {
      "get": {
        "tags": [
          "Anime"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Anime"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Anime"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Anime"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Anime"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Anime"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Anime"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Anime"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Anime"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Anime"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Anime"
                }
              }
            }
          }
        }
      }
    },
    "/api/Anime/{id}": {
      "get": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Anime"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Anime"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Anime"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Anime"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Anime"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Anime"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimePatch"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimePatch"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AnimePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Anime/{id}/Images": {
      "get": {
        "tags": [
          "Anime"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Image"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Image"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Image"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Image/{animeId}": {
      "post": {
        "tags": [
          "Image"
        ],
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ImageType"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          }
        }
      }
    },
    "/api/Image/{id}": {
      "get": {
        "tags": [
          "Image"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Image"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Anime": {
        "required": [
          "episodes",
          "titles"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "titles": {
            "minLength": 1,
            "type": "string"
          },
          "episodes": {
            "type": "integer",
            "format": "int32"
          },
          "genres": {
            "type": "string",
            "nullable": true
          },
          "studio": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AnimePatch": {
        "type": "object",
        "properties": {
          "titles": {
            "type": "string",
            "nullable": true
          },
          "episodes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "genres": {
            "type": "string",
            "nullable": true
          },
          "studio": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Image": {
        "required": [
          "animeId",
          "type",
          "url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "animeId": {
            "type": "integer",
            "format": "int64"
          },
          "url": {
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ImageType"
          }
        },
        "additionalProperties": false
      },
      "ImageType": {
        "enum": [
          "Poster",
          "Cover",
          "Screenshot"
        ],
        "type": "string"
      }
    }
  }
}